move grab related logic from game.js to grab-manager.js
