+++ game.js
@@ -21,13 +21,11 @@ import {
 import {getRenderer, sceneLowPriority, camera} from './renderer.js';
 import {
   downloadFile,
-  snapPosition,
   getDropUrl,
   handleDropJsonItem,
   makeId,
 } from './util.js';
 import {
-  maxGrabDistance,
   throwReleaseTime,
   throwAnimationDuration,
   walkSpeed,
@@ -45,53 +43,27 @@ import zTargeting from './z-targeting.js';
 import Avatar from './avatars/avatars.js';
 import {avatarManager} from './avatar-manager.js';
 import npcManager from './npc-manager.js';
+import grabManager from './grab-manager.js';
 
 const localVector = new THREE.Vector3();
 const localVector2 = new THREE.Vector3();
-const localVector3 = new THREE.Vector3();
-const localVector4 = new THREE.Vector3();
-const localVector5 = new THREE.Vector3();
-const localVector6 = new THREE.Vector3();
-const localVector7 = new THREE.Vector3();
-const localVector8 = new THREE.Vector3();
 // const localVector2D = new THREE.Vector2();
 const localQuaternion = new THREE.Quaternion();
-const localQuaternion2 = new THREE.Quaternion();
-const localQuaternion3 = new THREE.Quaternion();
 const localEuler = new THREE.Euler();
 const localMatrix = new THREE.Matrix4();
 const localMatrix2 = new THREE.Matrix4();
 const localMatrix3 = new THREE.Matrix4();
-const localBox = new THREE.Box3();
 const localRay = new THREE.Ray();
 
 //
 
 const physicsScene = physicsManager.getScene();
 let isMouseUp = false;
-const hand = 'right';
 
 // const zeroVector = new THREE.Vector3(0, 0, 0);
 // const oneVector = new THREE.Vector3(1, 1, 1);
 // const cubicBezier = easing(0, 1, 0, 1);
 // let redMesh = null;
-
-const _getGrabAction = i => {
-  const targetHand = i === 0 ? 'left' : 'right';
-  const localPlayer = playersManager.getLocalPlayer();
-  const grabAction = localPlayer.findAction(
-    action => action.type === 'grab' && action.hand === targetHand,
-  );
-  return grabAction;
-};
-const _getGrabbedObject = i => {
-  const grabAction = _getGrabAction(i);
-  const grabbedObjectInstanceId = grabAction?.instanceId;
-  const result = grabbedObjectInstanceId
-    ? metaversefileApi.getAppByInstanceId(grabbedObjectInstanceId)
-    : null;
-  return result;
-};
 
 const _unwearAppIfHasSitComponent = player => {
   const wearActions = player.getActionsByType('wear');
@@ -104,96 +76,6 @@ const _unwearAppIfHasSitComponent = player => {
     }
   }
 };
-
-const getPhysicalPosition = box => {
-  return localVector7.set(
-    (box.min.x + box.max.x) / 2,
-    box.min.y,
-    (box.min.z + box.max.z) / 2
-  );
-}
-
-// returns whether we actually snapped
-function updateGrabbedObject(
-  o,
-  grabMatrix,
-  offsetMatrix,
-  {collisionEnabled, handSnapEnabled, physx, gridSnap},
-) {
-  grabMatrix.decompose(localVector, localQuaternion, localVector2);
-  offsetMatrix.decompose(localVector3, localQuaternion2, localVector4);
-  const offset = localVector3.length();
-  localMatrix
-    .multiplyMatrices(grabMatrix, offsetMatrix)
-    .decompose(localVector5, localQuaternion3, localVector6);
-
-  let physicalOffset = null;
-  const physicsObjects = o.getPhysicsObjects();
-
-  // Compute physical local bounding box and it's position offset from app.position.
-  // THREE.Box3.getCenter() has a console error, so I calculate manually.
-  if(physicsObjects) {
-    localBox.makeEmpty();
-    for(const physicsObject of physicsObjects) {
-      const geometry = physicsObject.physicsMesh.geometry;
-      geometry.computeBoundingBox();
-      localBox.union(geometry.boundingBox);
-    }
-    physicalOffset = getPhysicalPosition(localBox);
-  }
-
-  // raycast from localPlayer in direction of camera angle
-  const collision = collisionEnabled && physicsScene.raycast(localVector, localQuaternion);
-
-  // raycast from grabbed object down perpendicularly
-  localQuaternion2.setFromAxisAngle(localVector2.set(1, 0, 0), -Math.PI * 0.5);
-  const downCollision = collisionEnabled && physicsScene.raycast(localVector5, localQuaternion2);
-
-  if (!!collision) {
-    const {point} = collision;
-    localVector6.fromArray(point);
-  }
-
-  if (!!downCollision) {
-    const {point} = downCollision;
-    localVector4.fromArray(point);
-  }
-
-  // Did the ray collide with any other object than the grabbed object? Need this check because on the first frame
-  // it collides with the grabbed object although physical actors are being disabled. This caused teleport issue.
-  const collNonGrabbedObj = !!collision && !o.physicsObjects.some(obj => obj.physicsId === collision.objectId);
-
-  // if collision point is closer to the player than the grab offset and collisionDown point
-  // is below collision point then place the object at collision point
-  if (collNonGrabbedObj && !!downCollision && localVector.distanceTo(localVector6) < offset && localVector4.y < localVector6.y) {
-    localVector5.copy(localVector6).sub(physicalOffset);
-  }
-
-  // if grabbed object would go below another object then place object at downCollision point
-  if (!!downCollision && localVector8.copy(localVector5).add(physicalOffset).y < localVector4.y) {
-    localVector5.setY(localVector4.sub(physicalOffset).y);
-  }
-
-  o.position.copy(localVector5);
-
-  const handSnap =
-    !handSnapEnabled ||
-    offset >= maxGrabDistance ||
-    !!collision ||
-    !!downCollision;
-  if (handSnap) {
-    snapPosition(o, gridSnap);
-    o.quaternion.setFromEuler(o.savedRotation);
-  } else {
-    o.quaternion.copy(localQuaternion3);
-  }
-
-  o.updateMatrixWorld();
-
-  return {
-    handSnap,
-  };
-}
 
 const _getCurrentGrabAnimation = () => {
   let currentAnimation = '';
@@ -479,11 +361,8 @@ highlightMesh.visible = false;
 sceneLowPriority.add(highlightMesh);
 let highlightedObject = null; */
 
-const highlightPhysicsMesh = _makeHighlightPhysicsMesh(buildMaterial);
-highlightPhysicsMesh.visible = false;
-sceneLowPriority.add(highlightPhysicsMesh);
-let highlightedPhysicsObject = null;
-let highlightedPhysicsId = 0;
+const grabHighlightPhysicsMesh = _makeHighlightPhysicsMesh(buildMaterial);
+grabManager.setHighlightPhysicsMesh(grabHighlightPhysicsMesh);
 
 const mouseHighlightPhysicsMesh = _makeHighlightPhysicsMesh(highlightMaterial);
 mouseHighlightPhysicsMesh.visible = false;
@@ -516,16 +395,7 @@ let mouseDomEquipmentHoverPhysicsId = 0;
 // let selectedLoadoutIndex = -1;
 
 const _delete = () => {
-  const grabbedObject = _getGrabbedObject(0);
-  if (grabbedObject) {
-    const localPlayer = playersManager.getLocalPlayer();
-    localPlayer.ungrab();
-
-    world.appManager.removeTrackedApp(grabbedObject.instanceId);
-  } else if (highlightedPhysicsObject) {
-    world.appManager.removeTrackedApp(highlightedPhysicsObject.instanceId);
-    highlightedPhysicsObject = null;
-  } else if (mouseSelectedObject) {
+  if (mouseSelectedObject) {
     world.appManager.removeTrackedApp(mouseSelectedObject.instanceId);
 
     if (mouseHoverObject === mouseSelectedObject) {
@@ -534,16 +404,7 @@ const _delete = () => {
     gameManager.setMouseSelectedObject(null);
   }
 };
-const _click = e => {
-  if (_getGrabbedObject(0)) {
-    const localPlayer = playersManager.getLocalPlayer();
-    localPlayer.ungrab();
-  } else {
-    if (highlightedPhysicsObject) {
-      _grab(highlightedPhysicsObject);
-    }
-  }
-};
+
 let lastUseIndex = 0;
 const _getNextUseIndex = animationCombo => {
   if (Array.isArray(animationCombo)) {
@@ -614,14 +475,6 @@ const _mouseup = () => {
   isMouseUp = true;
 };
 
-const _grab = object => {
-  const localPlayer = playersManager.getLocalPlayer();
-  localPlayer.grab(object);
-
-  gameManager.gridSnap = 0;
-  gameManager.editMode = false;
-};
-
 const hitRadius = 1;
 const hitHeight = 0.2;
 const hitHalfHeight = hitHeight * 0.5;
@@ -662,26 +515,14 @@ const _gameUpdate = (timestamp, timeDiff) => {
   const renderer = getRenderer();
   const localPlayer = playersManager.getLocalPlayer();
 
-  const _handlePush = () => {
-    if (gameManager.canPush()) {
-      if (ioManager.keys.forward) {
-        gameManager.menuPush(-1);
-      } else if (ioManager.keys.backward) {
-        gameManager.menuPush(1);
-      }
-    }
-  };
-  _handlePush();
-
-  const _updateGrab = () => {
-    const renderer = getRenderer();
+  const _updateGrabUseMesh = () => {
     const _isWear = o =>
       localPlayer.findAction(
         action => action.type === 'wear' && action.instanceId === o.instanceId,
       );
 
     grabUseMesh.visible = false;
-    if (!gameManager.editMode) {
+    if (!grabManager.editMode) {
       const avatarHeight = localPlayer.avatar ? localPlayer.avatar.height : 0;
       localVector
         .copy(localPlayer.position)
@@ -726,43 +567,8 @@ const _gameUpdate = (timestamp, timeDiff) => {
         }
       }
     }
-
-    for (let i = 0; i < 2; i++) {
-      const grabAction = _getGrabAction(i);
-      const grabbedObject = _getGrabbedObject(i);
-      if (grabbedObject && !_isWear(grabbedObject)) {
-        let position = null;
-        let quaternion = null;
-        if (renderer.xr.getSession()) {
-          const h = localPlayer[hand === 'left' ? 'leftHand' : 'rightHand'];
-          position = h.position;
-          quaternion = h.quaternion;
-        } else {
-          position = localVector2.copy(localPlayer.position);
-          quaternion = camera.quaternion;
-        }
-        localMatrix.compose(position, quaternion, localVector.set(1, 1, 1));
-
-        updateGrabbedObject(
-          grabbedObject,
-          localMatrix,
-          localMatrix3.fromArray(grabAction.matrix),
-          {
-            collisionEnabled: true,
-            handSnapEnabled: true,
-            physx,
-            gridSnap: gameManager.getGridSnap(),
-          },
-        );
-
-        grabUseMesh.setComponent(
-          'value',
-          localPlayer.actionInterpolants.activate.getNormalized(),
-        );
-      }
-    }
   };
-  _updateGrab();
+  _updateGrabUseMesh();
 
   zTargeting.update(timestamp, timeDiff);
 
@@ -826,59 +632,6 @@ const _gameUpdate = (timestamp, timeDiff) => {
     ioManager.setMovementEnabled(!pickUpAction);
   };
   _handlePickUp();
-
-  const _handlePhysicsHighlight = () => {
-    highlightedPhysicsObject = null;
-
-    if (gameManager.editMode) {
-      const {position, quaternion} = renderer.xr.getSession()
-        ? localPlayer.leftHand
-        : camera;
-      const collision = physicsScene.raycast(position, quaternion);
-      if (collision) {
-        const physicsId = collision.objectId;
-        highlightedPhysicsObject =
-          metaversefileApi.getAppByPhysicsId(physicsId);
-        highlightedPhysicsId = physicsId;
-      }
-    }
-  };
-  _handlePhysicsHighlight();
-
-  const _updatePhysicsHighlight = () => {
-    highlightPhysicsMesh.visible = false;
-
-    if (highlightedPhysicsObject) {
-      const physicsId = highlightedPhysicsId;
-
-      highlightedPhysicsObject.updateMatrixWorld();
-
-      const physicsObject =
-        metaversefileApi.getPhysicsObjectByPhysicsId(physicsId);
-      if (physicsObject) {
-        const {physicsMesh} = physicsObject;
-        highlightPhysicsMesh.geometry = physicsMesh.geometry;
-        highlightPhysicsMesh.matrixWorld
-          .copy(physicsMesh.matrixWorld)
-          .decompose(
-            highlightPhysicsMesh.position,
-            highlightPhysicsMesh.quaternion,
-            highlightPhysicsMesh.scale,
-          );
-
-        highlightPhysicsMesh.material.uniforms.uTime.value =
-          (now % 1500) / 1500;
-        highlightPhysicsMesh.material.uniforms.uTime.needsUpdate = true;
-        highlightPhysicsMesh.material.uniforms.uColor.value.setHex(
-          buildMaterial.uniforms.uColor.value.getHex(),
-        );
-        highlightPhysicsMesh.material.uniforms.uColor.needsUpdate = true;
-        highlightPhysicsMesh.visible = true;
-        highlightPhysicsMesh.updateMatrixWorld();
-      }
-    }
-  };
-  _updatePhysicsHighlight();
 
   const _updateMouseHighlight = () => {
     mouseHighlightPhysicsMesh.visible = false;
@@ -1289,7 +1042,7 @@ const _gameUpdate = (timestamp, timeDiff) => {
         cameraManager.getMode(),
       ) ||
         localPlayer.hasAction('aim')) &&
-      !_getGrabbedObject(0);
+      !grabManager.getGrabbedObject(0);
     crosshairEl.style.visibility = visible ? null : 'hidden';
   }
 
@@ -1321,8 +1074,6 @@ const _pushPlayerUpdates = () => {
   localPlayer.pushPlayerUpdates();
 };
 
-const rotationSnap = Math.PI / 6;
-
 /* const metaverseUi = {
   makeArrowLoader() {
     const app = metaversefileApi.createApp();
@@ -1342,7 +1093,7 @@ const _bindPointerLock = () => {
 
     gameManager.setMouseHoverObject(null);
     if (!pointerLockElement) {
-      gameManager.editMode = false;
+      grabManager.editMode = false;
     }
   });
 };
@@ -1375,8 +1126,6 @@ class GameManager extends EventTarget {
     super();
 
     this.menuOpen = 0;
-    this.gridSnap = 0;
-    this.editMode = false;
     // this.dragging = false;
     // this.draggingRight = false;
     this.contextMenu = false;
@@ -1431,10 +1180,6 @@ class GameManager extends EventTarget {
 
   menuDelete() {
     _delete();
-  }
-
-  menuClick(e) {
-    _click(e);
   }
 
   menuMouseDown() {
@@ -1553,16 +1298,6 @@ class GameManager extends EventTarget {
     );
   }
 
-  canRotate() {
-    return !!_getGrabbedObject(0);
-    // return !!world.appManager.grabbedObjects[0];
-  }
-
-  menuRotate(direction) {
-    const object = _getGrabbedObject(0);
-    object.savedRotation.y -= direction * rotationSnap;
-  }
-
   dropSelectedApp() {
     const app = loadoutManager.getSelectedApp();
     if (app) {
@@ -1588,51 +1323,10 @@ class GameManager extends EventTarget {
     }
   }
 
-  canPush() {
-    return !!_getGrabbedObject(0);
-    // return !!world.appManager.grabbedObjects[0] /*|| (editedObject && editedObject.isBuild)*/;
-  }
-
-  menuPush(direction) {
-    const localPlayer = playersManager.getLocalPlayer();
-    const grabAction = localPlayer.findAction(
-      action => action.type === 'grab' && action.hand === 'left',
-    );
-    if (grabAction) {
-      const matrix = localMatrix.fromArray(grabAction.matrix);
-      matrix.decompose(localVector, localQuaternion, localVector2);
-      localVector.z += direction * 0.1;
-      matrix
-        .compose(localVector, localQuaternion, localVector2)
-        .toArray(grabAction.matrix);
-    } else {
-      console.warn('trying to push with no grab object');
-    }
-  }
-
-  menuGridSnap() {
-    if (this.gridSnap === 0) {
-      this.gridSnap = 32;
-    } else if (this.gridSnap > 1) {
-      this.gridSnap /= 2;
-    } else {
-      this.gridSnap = 0;
-    }
-    // gridSnapEl.innerText = this.gridSnap > 0 ? (this.gridSnap + '') : 'off';
-  }
-
-  getGridSnap() {
-    if (this.gridSnap === 0) {
-      return 0;
-    } else {
-      return 4 / this.gridSnap;
-    }
-  }
-
   menuVDown() {
     const localPlayer = playersManager.getLocalPlayer();
-    if (_getGrabbedObject(0)) {
-      this.menuGridSnap();
+    if (grabManager.getGrabbedObject(0)) {
+      grabManager.menuGridSnap();
     } else {
       localPlayer.removeAction('dance');
 
@@ -1802,22 +1496,6 @@ class GameManager extends EventTarget {
 
   toggleAxis() {
     console.log('toggle axis');
-  }
-
-  async toggleEditMode() {
-    this.editMode = !this.editMode;
-    if (this.editMode) {
-      if (!cameraManager.pointerLockElement) {
-        await cameraManager.requestPointerLock();
-      }
-      if (this.mouseSelectedObject) {
-        this.setMouseSelectedObject(null);
-      }
-      if (_getGrabbedObject(0)) {
-        const localPlayer = playersManager.getLocalPlayer();
-        localPlayer.ungrab();
-      }
-    }
   }
 
   isJumping() {
